Master keep-alive timer working.
