+++ dnp3/examples/master-tokio-main.rs
@@ -26,17 +26,18 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     // Create the association
     let mut config = Configuration::default();
     config.auto_time_sync = Some(TimeSyncProcedure::LAN);
+    config.keep_alive_timeout = Some(Duration::from_secs(5));
     let mut association = master
         .add_association(EndpointAddress::from(1024)?, config, NullHandler::boxed())
         .await?;
 
     // Create event poll
-    let mut poll = association
-        .add_poll(
-            EventClasses::all().to_classes().to_request(),
-            Duration::from_secs(5),
-        )
-        .await?;
+    /*let mut poll = association
+    .add_poll(
+        EventClasses::all().to_classes().to_request(),
+        Duration::from_secs(5),
+    )
+    .await?;*/
 
     let mut reader = FramedRead::new(tokio::io::stdin(), LinesCodec::new());
 
@@ -88,7 +89,7 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
                     log::warn!("error: {}", err);
                 }
             }
-            "evt" => poll.demand().await,
+            //"evt" => poll.demand().await,
             "lts" => {
                 if let Err(err) = association.perform_time_sync(TimeSyncProcedure::LAN).await {
                     log::warn!("error: {}", err);
