+++ dnp3/src/master/tests/mod.rs
@@ -1,1 +1,3 @@
+mod harness;
+mod link_status;
 mod startup;
